+++ Cycle.NET.UnionDemo/Cycle.NET.UnionDemo.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net7.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="SdgApps.Common.DotnetSealedUnions" Version="1.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Cycle.NET\Cycle.NET.csproj" />
+  </ItemGroup>
+
+</Project>
