+++ Cycle.NET.sln
@@ -1,11 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26621.2
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33516.290
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cycle.NET", "Cycle.NET\Cycle.NET.csproj", "{6A0D9074-CE78-4217-A7D6-07554439A5BF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cycle.NET", "Cycle.NET\Cycle.NET.csproj", "{6A0D9074-CE78-4217-A7D6-07554439A5BF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cycle.NET.Demo", "Cycle.NET.Demo\Cycle.NET.Demo.csproj", "{E0B049E9-C963-4627-872B-56F6BCE5728C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cycle.NET.Demo", "Cycle.NET.Demo\Cycle.NET.Demo.csproj", "{E0B049E9-C963-4627-872B-56F6BCE5728C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cycle.NET.UnionDemo", "Cycle.NET.UnionDemo\Cycle.NET.UnionDemo.csproj", "{65F17018-1196-4153-A2D4-9711768A030B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{E0B049E9-C963-4627-872B-56F6BCE5728C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E0B049E9-C963-4627-872B-56F6BCE5728C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E0B049E9-C963-4627-872B-56F6BCE5728C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{65F17018-1196-4153-A2D4-9711768A030B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{65F17018-1196-4153-A2D4-9711768A030B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{65F17018-1196-4153-A2D4-9711768A030B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{65F17018-1196-4153-A2D4-9711768A030B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
